+++ src/components/common/auth/AuthButton.tsx
@@ -5,6 +5,7 @@ interface AuthButtonProps {
 const AuthButton = ({ children, disabled = false }: AuthButtonProps) => (
   <button
     type="submit"
+    disabled={disabled}
     className={`py-4 px-8 rounded-[6px] text-white text-lg font-semibold transition-opacity ${
       disabled ? 'bg-gray-400' : 'bg-blue-500 hover:opacity-90'
     }`}
