+++ src/pages/SignupPage.tsx
@@ -1,16 +1,12 @@
+import AuthLayout from '@/components/common/auth/AuthLayout';
 import LinkToLoginPage from '../components/signup/LinkToLoginPage';
 import SignupForm from '../components/signup/SignupForm';
 
 const SignupPage = () => (
-  <div>
-    <img
-      src="/assets/logo.svg"
-      alt="logo"
-      className="ml-auto mr-auto mt-[104px] mb-[40px]"
-    />
+  <AuthLayout>
     <SignupForm />
     <LinkToLoginPage />
-  </div>
+  </AuthLayout>
 );
 
 export default SignupPage;
